comprehensive tests
